Move divi compat to it's own file
