+++ projects/plugins/boost/compatibility/elementor.php
@@ -35,35 +35,3 @@ function exclude_elementor_library_custom_post_type( $post_types ) {
 
 add_filter( 'jetpack_boost_critical_css_post_types_singular', __NAMESPACE__ . '\exclude_elementor_library_custom_post_type' );
 add_filter( 'jetpack_boost_critical_css_post_types_archives', __NAMESPACE__ . '\exclude_elementor_library_custom_post_type' );
-
-/**
- * Defer JS can break Divi Builder.
- */
-function disable_defer_js_for_divi_builder( $should_defer_js ) {
-	$is_divi_builder = filter_input( INPUT_GET, 'et_fb', FILTER_VALIDATE_INT );
-
-	if ( 1 === (int) $is_divi_builder ) {
-		return false;
-	}
-
-	$is_divi_preview = filter_input(
-		INPUT_GET,
-		'et_pb_preview',
-		FILTER_VALIDATE_BOOLEAN,
-		array(
-			'flags' => FILTER_NULL_ON_FAILURE,
-		)
-	);
-
-	if ( true === $is_divi_preview ) {
-		return false;
-	}
-
-	if ( function_exists( 'is_et_pb_preview' ) && is_et_pb_preview() ) {
-		return false;
-	}
-
-	return $should_defer_js;
-}
-
-add_filter( 'jetpack_boost_should_defer_js', __NAMESPACE__ . '\disable_defer_js_for_divi_builder' );
